requirements.txt
annotated-types==0.7.0
anyio==4.6.0
boto3==1.35.34
botocore==1.35.34
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
mangum==0.19.0
pybreaker==1.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
redis==5.1.1
s3transfer==0.10.2
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
